+++ navigator-compose/navigator-compose-hilt/src/androidTest/java/com/kpstv/navigation/compose/hilt/internals/App.kt
@@ -1,0 +1,7 @@
+package com.kpstv.navigation.compose.hilt.internals
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+public class App : Application()