+++ navigator-compose/samples/basic-sample/src/main/java/com/kpstv/navigation/compose/sample/ui/viewmodels/HiltTestViewModel.kt
@@ -1,0 +1,25 @@
+package com.kpstv.navigation.compose.sample.ui.viewmodels
+
+import androidx.lifecycle.SavedStateHandle
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+import kotlin.random.Random
+
+@HiltViewModel
+class HiltTestViewModel @Inject constructor(private val savedStateHandle: SavedStateHandle) : ViewModel() {
+    var data: String = "NOT_SET"
+        private set
+
+    init {
+        savedStateHandle.get<String>("data")?.also { data = it }
+        if (data == "NOT_SET") {
+            data = "%.3f".format(Random.nextDouble())
+        }
+    }
+
+    fun set(data: String) {
+        savedStateHandle.set("data", data)
+        this.data = data
+    }
+}